pyproject.toml
[project]
name = "blueprint"
version = "0.1.0"
description = "Configuration files for my Mac"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
]
